Create README.md
This is a program that will run on Windows and control your Roku Device(s). So far, I have only tested it on the Roku Streaming Stick but it seems to work fine. It cannot automatically locate the Steaming Stick so you will have to know the IP Address. I'm pretty sure that finding the IP address of the device wouldn't be too hard, I just haven't figured that out yet.